requirements.txt
pytest

PyYAML~=6.0